metal: Significant network refactoring
This commit does significant network refactoring.  Among other things
it does the following:
  * Moves DHCP in to this repo with dnsmasq
  * Moves the dnsmasq configuration to a file
  * Makes shoelaces mappings IP based as well as DNS based
  * Adjusts certain PXE parameters

This change is motivated by the replacement of the router that was
running the minicluster from a SOHO router to a dedicated Juniper rack
mount router.  The Juniper box is easier to integrate into the rest of
the network, but does not have these services onboard.

The Shoelaces change is motivated by the dns based mapping being
generally unreliable.  In order to not have this unreliable component
as part of the critical path, mappings for IP are also provided to
provide two possible paths for shoelaces to be able to bootstrap
machines with.